fix: use u32 for block height to avoid hex outputs in block subjects
Fuel BlockHeight type `to_string()` outputs hexadecimal which
is optimized and perfect for their use-case. For our streams,
it might be better to store the integer value directly since it will
save us a tidbit compute time when decoding, which will happen alot.

Because of the hexadecimal output, when decoding the block height's
subject, it failed as it expecting a regular integer, not an hexadecimal.
This change fixes this by replacing the `BlockHeight` type with a `u32`.