fix: publish compact json rather than pretty json
Currently, we are publishing our subjects and payloads using pretty json.
This quickly added up and caused a large ROM consumption in our NATS
cluster. For obvious reasons, pretty json is way heavier compared with it's
compact equivalent. This change fixes this by only publishing the blocks
and transactions serialized to compact json.