refactor: ensure Stream nomenclature is consistent
This change ensures the ambiguity between `fuel-streams-core` stream and `fuel-streams` stream
is contained in the `fuel-streams` crate. In there, we will have to explicitly manage the ambiguity by
accessing the internal stream with the `fuel_streams_core` namespace.

Also, to prevent calling `stream.stream()` in the tests, this commit renames the `stream` field to `internal_stream`
to further curb the ambiguity.